testing suite works, and fixxed sv compile issue

+++ crates/backend/src/fpga/mod.rs
@@ -61,10 +61,6 @@ impl JITBackend for FPGABackend {
     ) {
         //println!("{:?}", &graph);
 
-        let path = format!("{}/{}", plot, name);
-        self.link.name = name;
-        self.config = config.unwrap();
-        self.path = path;
         for nodeid in graph.node_indices() {
             let node = &graph[nodeid];
             if let Some((pos, blockid)) = node.block {
@@ -72,18 +68,30 @@ impl JITBackend for FPGABackend {
                 self.link.add_block(block, pos);
             }
         }
-        println!("generating link file");
-        self.link.generate_link_file(Path::new(&format!("FPGA/bin/{}/link.json", self.path)));
-        println!("generating veruilog");
-        assembler::generate_verilog(&graph, Path::new(&format!("FPGA/bin/{}/redstone.sv", self.path)));
-        println!("create_project");
-        self.config.create_project(Path::new(&format!("FPGA/bin/{}/prj/prj.tcl",self.path)), self.link.output_bits as u32, self.link.input_bits);
-        println!("compile");
-        self.config.compile(Path::new(&format!("FPGA/bin/{}/prj", self.path)));
-        println!("done");
-        _ = copy(Path::new(&format!("FPGA/bin/{}/prj/RoC.sof", self.path)), Path::new(&format!("FPGA/bin/{}/RoC.sof", self.path)));
-        if !options.compile_verilog {
-            _ = remove_dir_all(Path::new(&format!("FPGA/bin/{}/prj", self.path)));  
+
+        if !options.sim_test {
+
+            let path = format!("{}/{}", plot, name);
+            self.link.name = name;
+            self.config = config.unwrap();
+            self.path = path;
+
+            println!("generating link file");
+            self.link.generate_link_file(Path::new(&format!("FPGA/bin/{}/link.json", self.path)));
+            println!("generating veruilog");
+            assembler::generate_verilog(&graph, Path::new(&format!("FPGA/bin/{}/redstone.sv", self.path)));
+            println!("create_project");
+            self.config.create_project(Path::new(&format!("FPGA/bin/{}/prj/prj.tcl",self.path)), self.link.output_bits as u32, self.link.input_bits);
+            println!("compile");
+            self.config.compile(Path::new(&format!("FPGA/bin/{}/prj", self.path)));
+            println!("done");
+            _ = copy(Path::new(&format!("FPGA/bin/{}/prj/RoC.sof", self.path)), Path::new(&format!("FPGA/bin/{}/RoC.sof", self.path)));
+            if !options.compile_verilog {
+                _ = remove_dir_all(Path::new(&format!("FPGA/bin/{}/prj", self.path)));  
+            }
+        }
+        else {
+            assembler::generate_verilog(&graph, Path::new(&format!("{name}/roc.sv")));
         }
     }
